Changes usings in Functions to match utils namespaces
