+++ Function/GetEmissionHistory/GetEmissionHistoryTest.cs
