+++ Function/GetEmissionHistory/Utils/ConvertToEmptyEmission.cs
