+++ Function/GetEmissionHistory/Utils/DailyEmission.cs
