+++ Function/GetEmissionStatistics/GetEmissionStatisticsTest.cs
