+++ Function/GetEmissionStatistics/Utils/EmissionStatistic.cs
