+++ Function/GetEmissionStatistics/Utils/StatisticsCalculatorTest.cs
