+++ Function/GetUserActivities/Utils/AllActivityRetrieval.cs
