+++ Function/GetUserActivities/Utils/DataRetrievalTest.cs
