+++ Function/GetUserActivities/Utils/ReturnedInfo.cs
