+++ Function/SubmitUserActivities/Utils/DailyEmissionsUpdate.cs
