+++ Function/SubmitUserActivities/Utils/DeleteUserActivities.cs
