+++ Function/SubmitUserActivities/Utils/EmissionCalculator.cs
