+++ Function/SubmitUserActivities/Utils/EmissionCalculatorTest.cs
