+++ Function/SubmitUserActivities/Utils/InputActivity.cs
