+++ Function/SubmitUserActivities/Utils/SubmitUserActivitiesBody.cs
